Split FixturesTrait into JoindIn & Raffle FixtureTraits
